feat: refactor and correct

+++ ms_invoice_management/infrastructure/driven-adapters/jpa-repository/src/main/java/co/com/management/jpa/helper/Util.java
@@ -50,7 +50,7 @@ public class Util {
         return Map.of(
                 "p_invoice_code", invoice.getId(),
                 "p_create_date", Timestamp.valueOf(invoice.getCreatedDate()),
-                "p_total_amount", BigDecimal.valueOf(invoice.getTotalAmount()),
+                "p_total_amount", invoice.getTotalAmount(),
                 "p_client_id", client.getId(),
                 "p_insert_invoice", client.getState() ? 1 : 0,
                 "p_products_code", new OracleHelper(arrays.getVarchar2List(), productData.getCodes()),
