+++ ms_invoice_management/infrastructure/driven-adapters/rest-consumer/src/main/java/co/com/management/consumer/mapper/RMapper.java
@@ -29,7 +29,7 @@ public class RMapper {
                 .id(invoiceData.getCode())
                 .products(products)
                 .clientId(clientId)
-                .totalAmount(dto.getData().getTotalAmount().doubleValue())
+                .totalAmount(dto.getData().getTotalAmount())
                 .build();
 
     }
